+++ plugin.php
@@ -1,8 +1,8 @@
 <?php
 
-$id                 = "pchfp";
-$version            = "3.1.0";
-$ilias_min_version  = "6.0";
+$id = "pchfp";
+$version = "3.1.1";
+$ilias_min_version = "6.0";
 $ilias_max_version  = "7.999";
 $responsible        = "sr solutions ag";
 $responsible_mail   = "support@sr.solutions";